refactor: migrate vm0 preference to vm0 zero preference
Migrate the preference CLI command from old /api/user/preferences
route to existing /api/zero/user-preferences route. Move shared
schemas into zero contract, create new zero API client, and clean
up all old artifacts (command, route, contract, API client).

Closes #6429
