+++ turbo/apps/cli/src/commands/zero/index.ts
@@ -2,6 +2,7 @@ import { Command } from "commander";
 import { zeroOrgCommand } from "./org";
 import { agentCommand } from "./agent";
 import { zeroConnectorCommand } from "./connector";
+import { zeroPreferenceCommand } from "./preference";
 import { zeroScheduleCommand } from "./schedule";
 import { zeroSecretCommand } from "./secret";
 import { zeroVariableCommand } from "./variable";
@@ -11,6 +12,7 @@ export const zeroCommand = new Command("zero")
   .addCommand(zeroOrgCommand)
   .addCommand(agentCommand)
   .addCommand(zeroConnectorCommand)
+  .addCommand(zeroPreferenceCommand)
   .addCommand(zeroScheduleCommand)
   .addCommand(zeroSecretCommand)
   .addCommand(zeroVariableCommand);
