+++ turbo/apps/cli/src/index.ts
@@ -15,7 +15,6 @@ import { logsCommand } from "./commands/logs";
 
 import { initCommand } from "./commands/init";
 
-import { preferenceCommand } from "./commands/preference";
 import { upgradeCommand } from "./commands/upgrade";
 import { whoamiCommand } from "./commands/whoami";
 import { zeroCommand } from "./commands/zero";
@@ -41,7 +40,6 @@ program.addCommand(cookCommand);
 program.addCommand(logsCommand);
 
 program.addCommand(initCommand);
-program.addCommand(preferenceCommand);
 program.addCommand(upgradeCommand);
 program.addCommand(whoamiCommand);
 program.addCommand(zeroCommand);
