--- turbo/apps/web/app/api/user/preferences/route.ts
@@ -1,93 +1,0 @@
-/**
- * Backward-compatible route for /api/user/preferences.
- *
- * The CLI still uses the old contract (userPreferencesContract) with
- * GET and PUT methods at this path. This route delegates to the shared
- * service layer (same as /api/zero/user-preferences).
- *
- * Will be removed once the CLI contracts are migrated.
- */
-import {
-  createHandler,
-  createSafeErrorHandler,
-  tsr,
-} from "../../../../src/lib/ts-rest-handler";
-import { userPreferencesContract, createErrorResponse } from "@vm0/core";
-import { initServices } from "../../../../src/lib/init-services";
-import {
-  requireAuth,
-  isAuthError,
-} from "../../../../src/lib/auth/require-auth";
-import { resolveOrg } from "../../../../src/lib/org/resolve-org";
-import {
-  getUserPreferences,
-  updateUserPreferences,
-} from "../../../../src/lib/user/user-preferences-service";
-import { isBadRequest } from "../../../../src/lib/errors";
-
-const router = tsr.router(userPreferencesContract, {
-  get: async ({ headers }, { request }) => {
-    initServices();
-
-    const authCtx = await requireAuth(headers.authorization);
-    if (isAuthError(authCtx)) return authCtx;
-
-    const orgSlug = new URL(request.url).searchParams.get("org");
-    const { org } = await resolveOrg(authCtx, orgSlug);
-
-    const prefs = await getUserPreferences(org.orgId, authCtx.userId);
-
-    return {
-      status: 200 as const,
-      body: {
-        timezone: prefs.timezone,
-        notifyEmail: prefs.notifyEmail,
-        notifySlack: prefs.notifySlack,
-        pinnedAgentIds: prefs.pinnedAgentIds,
-        sendMode: prefs.sendMode,
-      },
-    };
-  },
-
-  update: async ({ body, headers }, { request }) => {
-    initServices();
-
-    const authCtx = await requireAuth(headers.authorization);
-    if (isAuthError(authCtx)) return authCtx;
-
-    const orgSlug = new URL(request.url).searchParams.get("org");
-    const { org } = await resolveOrg(authCtx, orgSlug);
-
-    try {
-      const prefs = await updateUserPreferences(org.orgId, authCtx.userId, {
-        timezone: body.timezone,
-        notifyEmail: body.notifyEmail,
-        notifySlack: body.notifySlack,
-        pinnedAgentIds: body.pinnedAgentIds,
-        sendMode: body.sendMode,
-      });
-
-      return {
-        status: 200 as const,
-        body: {
-          timezone: prefs.timezone,
-          notifyEmail: prefs.notifyEmail,
-          notifySlack: prefs.notifySlack,
-          pinnedAgentIds: prefs.pinnedAgentIds,
-          sendMode: prefs.sendMode,
-        },
-      };
-    } catch (error) {
-      if (isBadRequest(error)) {
-        return createErrorResponse("BAD_REQUEST", "Invalid request");
-      }
-      throw error;
-    }
-  },
-});
-
-const handler = createHandler(userPreferencesContract, router, {
-  errorHandler: createSafeErrorHandler("user-preferences-compat"),
-});
-
-export { handler as GET, handler as PUT };
